requirements.txt
faiss-cpu
google-generativeai
gptcache
huggingface-hub
langchain
langchain-community
langchain-openai==0.1.8
langchain-google-genai==1.0.6
llama-index==0.9.39
openai==1.32.0
opencv-python
orjson
pandas
pdf2image
pillow
pypdf
pytesseract
rank-bm25
regex
requests
sentence-transformers
SQLAlchemy

youtube-search-python

PyPDF2
python-debian

unstructured
openpyxl
streamlit
torch